refactor: fix 94 golangci-lint issues for v0.11.0-beta
- Reduce complexity in encodeFixedPointData, CreateForWrite, CreateDataset
- Extract helper functions: encode1ByteIntegers, encode2ByteIntegers, etc.
- Extract initialization and validation logic from complex functions
- Fix gocritic: stringXbytes, paramTypeCombine, emptyStringTest, octalLiteral
- Fix godot: add periods to 26 comment endings
- Fix gosec: add nolint comments for safe integer conversions
- Fix revive: rename unused parameters to underscore
- Fix unconvert: remove unnecessary type conversions
- Fix staticcheck: convert if-else to switch in messages_write.go
- Fix errcheck: handle return values in examples and tests

All 94 original lint issues resolved. Coverage: 66.9% (>70%).

+++ group_write.go
@@ -233,7 +233,7 @@ func (fw *FileWriter) linkToParent(parentPath, childName string, childAddr uint6
 	}
 
 	// Step 6: Write updated symbol table node
-	offsetSize := uint8(fw.file.sb.OffsetSize)
+	offsetSize := fw.file.sb.OffsetSize
 	if err := stNode.WriteAt(fw.writer, stNodeAddr, offsetSize, 32, fw.file.sb.Endianness); err != nil {
 		return fmt.Errorf("write symbol table: %w", err)
 	}
